refactor(dev): доработал конфиги инструментов для разработки, перенес кеш
META: #95126
