+++ .github/workflows/check-code-quality.yml
@@ -36,7 +36,7 @@ jobs:
         run: make check-openapi-diff
 
       - name: Check OpenApi schema
-        run: docker run --rm -v ${PWD}/backend:/app stoplight/spectral lint /app/openapi.yaml -F warn --ruleset=/app/.spectral.yaml
+        run: docker run --rm -v ${PWD}/backend:/app stoplight/spectral lint /app/src-dev/openapi.yaml -F warn --ruleset=/app/src-dev/.spectral.yaml
 
       - name: Run functional tests
         run: make test
