+++ backend/src-dev/Rector/rector.config.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-use Dev\Rector\OneFlushInClassRector;
-use Dev\Rector\RequestMethodInsteadOfStringRector;
-use Dev\Rector\ResolversInActionRector;
+use Dev\Rector\Rules\OneFlushInClassRector;
+use Dev\Rector\Rules\RequestMethodInsteadOfStringRector;
+use Dev\Rector\Rules\ResolversInActionRector;
 use Rector\CodeQuality\Rector\Class_\InlineConstructorDefaultToPropertyRector;
 use Rector\CodeQuality\Rector\Identical\FlipTypeControlToUseExclusiveTypeRector;
 use Rector\Config\RectorConfig;
@@ -21,12 +21,13 @@ use Rector\Symfony\Symfony43\Rector\MethodCall\WebTestCaseAssertResponseCodeRect
 use Rector\TypeDeclaration\Rector\FunctionLike\AddParamTypeSplFixedArrayRector;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->cacheDirectory('/app/var/cache/rector');
+    $rectorConfig->cacheDirectory(__DIR__.'/../cache/rector');
+
     $rectorConfig->paths([
-        '/app/src',
-        '/app/src-dev',
-        '/app/migrations',
-        '/app/tests',
+        __DIR__.'/../../src',
+        __DIR__.'/../',
+        __DIR__.'/../../migrations',
+        __DIR__.'/../../tests',
     ]);
 
     $rectorConfig->parallel(seconds: 360);
@@ -48,16 +49,17 @@ return static function (RectorConfig $rectorConfig): void {
     ]);
 
     $rectorConfig->skip([
-        ClassPropertyAssignToConstructorPromotionRector::class => ['/app/src/*/Domain/*'],
+        ClassPropertyAssignToConstructorPromotionRector::class => [dirname(__DIR__, 2).'/src/*/Domain/*'],
         InlineConstructorDefaultToPropertyRector::class,
         FlipTypeControlToUseExclusiveTypeRector::class,
-        RemoveUnusedPrivatePropertyRector::class => ['/app/src/*/Domain/*'],
+        RemoveUnusedPrivatePropertyRector::class => [dirname(__DIR__, 2).'/src/*/Domain/*'],
         WebTestCaseAssertIsSuccessfulRector::class,
         WebTestCaseAssertResponseCodeRector::class,
         AddParamTypeSplFixedArrayRector::class => [
-            __DIR__.'/PHPCsFixer',
+            __DIR__.'/../PHPCsFixer',
         ],
-        __DIR__.'/Maker/Resources/skeleton',
+        __DIR__.'/../Maker/Resources/skeleton',
+        __DIR__.'/../cache',
     ]);
 
     $rectorConfig->rule(RequestMethodInsteadOfStringRector::class);
