--- backend/src-dev/phpstan-config.neon
@@ -1,24 +1,0 @@
-parameters:
-    scanDirectories:
-        - ../var/cache
-    checkUninitializedProperties: true
-    level: 9
-    tmpDir: ../var/cache/phpstan
-    paths:
-        - ../src
-        - ../src-dev
-        - ../tests
-    excludePaths:
-        - ./Maker/Vendor/EntityGenerator.php
-    doctrine:
-        objectManagerLoader: ./PHPStan/object-manager.php
-
-includes:
-    - ../vendor/phpstan/phpstan/conf/bleedingEdge.neon
-    - ../vendor/phpstan/phpstan-doctrine/extension.neon
-    - ../vendor/phpstan/phpstan-doctrine/rules.neon
-    - ../vendor/phpstan/phpstan-strict-rules/rules.neon
-    - ./PHPStan/phpstan-baseline.neon
-
-rules:
-    - PHPStan\Rules\Doctrine\ORM\EntityNotFinalRule
