+++ backend/src-dev/phpunit.xml
@@ -6,15 +6,13 @@
          backupGlobals="false"
          colors="true"
          bootstrap="../tests/bootstrap.php"
-         cacheResultFile="../www/var/cache/.phpunit.result.cache"
+         cacheDirectory="./cache"
 >
     <php>
         <ini name="display_errors" value="1" />
         <ini name="error_reporting" value="-1" />
         <server name="APP_ENV" value="test" force="true" />
         <server name="SHELL_VERBOSITY" value="-1" />
-        <server name="SYMFONY_PHPUNIT_REMOVE" value="" />
-        <server name="SYMFONY_PHPUNIT_VERSION" value="9.5" />
         <server name="APP_DEBUG" value="false" />
     </php>
