+++ backend/src-dev/psalm.xml
@@ -6,7 +6,7 @@
     xsi:schemaLocation="https://getpsalm.org/schema/config ../vendor/vimeo/psalm/config.xsd"
     findUnusedBaselineEntry="true"
     findUnusedCode="false"
-    cacheDirectory="../var/cache/psalm"
+    cacheDirectory="./cache/psalm"
 >
     <projectFiles>
         <directory name="../src" />
@@ -14,6 +14,7 @@
         <directory name="../tests" />
         <ignoreFiles>
             <directory name="../vendor" />
+            <directory name="./cache" />
             <file name="../src/*/config.php" />
             <file name="../src/*/*/config.php" />
             <file name="../src/Infrastructure/Kernel.php" />
@@ -44,7 +45,7 @@
 
         <UnusedVariable>
             <errorLevel type="suppress">
-                <file name="./Rector/OneFlushInClassRector.php" />
+                <file name="./Rector/Rules/OneFlushInClassRector.php" />
             </errorLevel>
         </UnusedVariable>
         <PluginIssue name="IssueNameEmittedByPlugin" errorLevel="info" />
